Removes stop words from a text file
Doesn't work quite yet. I had a brain freeze and can't remember how to change a list while iterating over it. Or should I say I know you shouldn't but I can't think of a clever way to do it that will scale up to the amount of data that it will be given.